[core] Keep run merge partition ordering local
